# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Lapa Casa Hostel - Channel Manager API
  description: |
    Documentación de la API del sistema de reservas y channel manager de
    Lapa Casa Hostel.

    ## Autenticación
    Las rutas bajo /admin/* requieren un JWT (cookie httpOnly `lch_admin`
    o header `Authorization: Bearer <token>`), obtenido via
    POST /admin/login. El resto de las rutas son públicas.

    ## Rate limiting
    Cada grupo de rutas tiene su propio límite (ver rate-limit.md o
    routes/index.ts) -- desde 3 req/seg en /bookings hasta 120 req/min
    en /availability.
  version: 1.0.0
  contact:
    name: Lapa Casa Hostel Tech Team
    email: lapalandiarj@gmail.com
    url: https://lapacasario.com
  license:
    name: Proprietary
servers:
  - url: https://api.lapacasario.com/api/v1
    description: Producción (aún no desplegado bajo este dominio -- ver sección 11 de la auditoría de 17 secciones)
  - url: http://localhost:3001/api/v1
    description: Desarrollo local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: También aceptado como cookie httpOnly `lch_admin`
    ownerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        JWT de administrador de apartamento (role owner), separado del admin
        único de arriba. También aceptado como cookie httpOnly `lch_owner`.
...
paths:
...
  /webhooks/booking:
    post:
      tags: [Webhooks]
      summary: Webhook de Booking.com
      description: Requiere header X-Api-Key y, según configuración, X-Signature + allowlist de IPs.
      security: []
      responses:
        '200': { description: Procesado }
        '401': { description: API key/firma inválida }
  /webhooks/expedia:
    post:
      tags: [Webhooks]
      summary: Webhook de Expedia
      description: Mismo esquema de auth que /webhooks/booking, secrets propios (EXPEDIA_WEBHOOK_*).
      security: []
      responses:
        '200': { description: Procesado }
        '401': { description: API key/firma inválida }

  # ── Admin Auth ─────────────────────────────────────────────────────────
...